Refactor: Sections Wrapper클래스 생성
