+++ src/test/java/nextstep/subway/line/domain/LineTest.java
@@ -1,0 +1,7 @@
+package nextstep.subway.line.domain;
+
+import static org.junit.jupiter.api.Assertions.*;
+
+class LineTest {
+
+}
